Recreate: Leveling logic
Remade leveling logic, so that classes are now separated in files. This might be a temporary change, so that the project can go further quicker.

This also enables the possibility of including functions and other things that are exclusive to a class inside its logic, that is an interesting tradeoff for easy class creation in the interfaces.

All of this will be considered, but this version will be used for now.