Update: Spell Choice Modal
Finished the main idea for the spell choice modal, when clicking the special input for the choice, a window opens listing all spells for the options sent to the modal.

There the user can see the descriptions and select an amount of spells also determined by the options.

When finished the name of the spells is joined into a string that is passed to the input initially clicked, listing the name of the spells chosen.