Create: Console.js, Screen (Log)
console class to override default JS console, since that does not run on the backend.

To accompany that, a log window is now shown in the screen, that can be minimized by the user. This window only shows text from console.log on the backend that explicitly targets the type of user to see it.

So public logs will be also shown to non GMs.