Fix: X, Y, Distance calculation for larger token sizes.
Since maptool returns X and Y of the leftmost square of a token instead of its center, all distance validations were incorrect for characters of such sizes.

To fix it, the X and Y returned by maptool are now offset to the actual center of the creature, the same process is done when setting a new value for them.

Additionally, calculate_distance() now reduces the distance based on the distance of a token's edge and its center. (So that it always returns distance from the edge)