Update: Duration for Conditions
Now instead of having a duration be a countdown that is decreased for every condition on a creature on its turn start, we instead use an END_TIME attribute on the condition, that is tied to the global time set by the new Time class, and use a few new methods to calculate the remaining duration for the interfaces.

On the time class, a new check_expired_conditions method has been added so that the conditions expiration is triggered by time advancement for all Creatures on the current map.