feat (creature spells): add support for innate spells
Innate spells are spells that can be cast without spending spell slots. This has been done specially for monsters, although some classes get a form of at later levels (unimplemented)

This change only adds the structure to support it, although it should be easy to implement the costless attribute on top of the current spellcasting system later.